move rkey check to prepareCreate
